+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## [0.0.6] fix  The method 'cancel' was called on null
+
+## [0.0.5] add extra like count
+
+添加中文文档
+
 ## [0.0.4] add on Tap delay
 
 ## [0.0.3] add on Tap callback